Convert data function into class
This will make it much easier to coordinate different data sets
and base them on the same match schedule.
