+++ candystore/__init__.py
@@ -1,3 +1,3 @@
 """Factories for randomised AFL data sets for testing purposes."""
 
-from .fixtures import generate_fixtures
+from .afl_data_factory import AFLDataFactory
